refactor(cdk-experimental/menu): move listeners from host object to host listener decorator
Some of the Cdk directives are expected to be extended by some subclasses (e.g. material). In Ivy
the host metadata will be merged whereas in ViewEngine it will be overridden forcing duplication of
listeners in the host object on the subclass which results in double listeners in Ivy. Moving to
@HostListener prevents having to duplicate them in subclasses.